Refactor authentication DTOs to use KeycloakTokenResponseDto and update related services
