--- services/api/Api/DTOs/Auth/AuthResponse.cs
@@ -1,12 +1,0 @@
-namespace Api.DTOs.Auth;
-
-/// <summary>
-/// Standard authentication response containing tokens and user information.
-/// </summary>
-public class AuthResponse
-{
-    public string Token { get; set; } = string.Empty;
-    public string RefreshToken { get; set; } = string.Empty;
-    public DateTime ExpiresAt { get; set; }
-    public UserDto User { get; set; } = new();
-}
