--- services/api/Api/DTOs/Auth/LoginRequest.cs
@@ -1,16 +1,0 @@
-namespace Api.DTOs.Auth;
-
-using System.ComponentModel.DataAnnotations;
-
-/// <summary>
-/// User login request.
-/// </summary>
-public class LoginRequest
-{
-    [Required]
-    [EmailAddress]
-    public string Email { get; set; } = string.Empty;
-
-    [Required]
-    public string Password { get; set; } = string.Empty;
-}
