+++ services/api/Api/DTOs/Auth/LoginRequestDto.cs
@@ -1,0 +1,10 @@
+namespace Api.DTOs.Auth;
+
+/// <summary>
+/// Login request.
+/// </summary>
+public class LoginRequestDto
+{
+    public string Username { get; set; } = string.Empty;
+    public string Password { get; set; } = string.Empty;
+}
