+++ services/api/Api/DTOs/Auth/LogoutRequestDto.cs
@@ -1,0 +1,9 @@
+namespace Api.DTOs.Auth;
+
+/// <summary>
+/// Logout request.
+/// </summary>
+public class LogoutRequestDto
+{
+    public string RefreshToken { get; set; } = string.Empty;
+}
