+++ services/api/Api/DTOs/Auth/RefreshTokenRequestDto.cs
@@ -1,0 +1,9 @@
+namespace Api.DTOs.Auth;
+
+/// <summary>
+/// Token refresh request.
+/// </summary>
+public class RefreshTokenRequestDto
+{
+    public string RefreshToken { get; set; } = string.Empty;
+}
