--- services/api/Api/DTOs/UserDto.cs
@@ -1,18 +1,0 @@
-namespace Api.DTOs;
-
-/// <summary>
-/// User information returned to the client.
-/// </summary>
-public class UserDto
-{
-    public string Id { get; set; } = string.Empty;
-    public string Email { get; set; } = string.Empty;
-    public string? UserName { get; set; }
-    public string? FirstName { get; set; }
-    public string? LastName { get; set; }
-    public bool EmailConfirmed { get; set; }
-    public bool IsActive { get; set; }
-    public DateTime CreatedAt { get; set; }
-    public DateTime? LastLoginAt { get; set; }
-    public List<string> Roles { get; set; } = new();
-}
