+++ services/api/Api/Program.cs
@@ -6,7 +6,6 @@ using Api.Services.JwtTokenValidation;
 using Api.Services.Keycloak;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.IdentityModel.Tokens;
