--- services/api/Api/services/Auth/IAuthService.cs
@@ -1,12 +1,0 @@
-namespace Api.Services.Auth;
-
-using Api.DTOs.Auth;
-
-public interface IAuthService
-{
-    Task<AuthResponse> RegisterAsync(RegisterRequest request);
-    Task<AuthResponse> LoginAsync(LoginRequest request);
-    Task<AuthResponse> RefreshTokenAsync(string refreshToken);
-    Task<bool> RevokeTokenAsync(string token);
-    Task SyncWithKeycloakAsync(string userId);
-}
